+++ lib/kumi/evaluation_wrapper.rb
@@ -1,0 +1,20 @@
+module Kumi
+  EvaluationWrapper = Struct.new(:ctx) do
+    def initialize(ctx)
+      @ctx = ctx
+      @__schema_cache__ = {} # memoization cache for bindings
+    end
+
+    def [](key)
+      @ctx[key]
+    end
+
+    def keys
+      @ctx.keys
+    end
+
+    def key?(key)
+      @ctx.key?(key)
+    end
+  end
+end